update libft
- allocation:
use ft_malloc instead of malloc so it can be set to exit on allocation
failure.
- ft_realloc:
do not free the addr arg.
- vector:
free the old addr in ft_realloc_vector, cause ft_realloc does not free.
make sure that capacity is always >= 1.
- ft_strjoin:
return NULL if either one of the args is NULL.
- ft_strings_array_length:
add a function to calculate the length of a NULL terminated array of
strings.